requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
pyspellchecker
